Report for Employee and picks done... Check it out
I decided to add a text box where the user will enter a number, which
will be the id that he/she wishes to see.  Once the "Update Report"
button is clicked it will show all the picks the employee had, and will
also show a bar graph showing the total weight the employee picked per
field.